requirements-backend.txt
gradio==4.36.1
requests==2.31.0
beautifulsoup4==4.12.2
pillow==10.1.0
urllib3==2.1.0
bing-image-downloader==1.3.11
yandex-reverse-image-search==1.0.0
lxml==4.9.3
python-dotenv==1.0.0
opencv-python==4.8.1.78
exifread==3.0.0
geopy==2.4.0
httpx==0.25.2
aiofiles==23.2.1
fake-useragent==1.4.0
google-search-results==2.4.2

# Backend & API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1

# Database
psycopg2-binary==2.9.9
pymysql==1.1.0

# Caching & Queue
redis==5.0.1
celery==5.3.4
flower==2.0.1

# Authentication & Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
pyjwt==2.8.1
cryptography==41.0.7

# Async
aiohttp==3.9.1
asyncio==3.4.3

# Monitoring & Logging
python-json-logger==2.0.7
sentry-sdk==1.38.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1
ipython==8.18.1
